requirements.txt
RPi.GPIO>=0.7.1
requests>=2.28.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
